requirements.txt
web3~=6.4.0
python-dotenv~=1.0.0
hexbytes~=0.3.0
aiohttp~=3.8.4
requests~=2.26.0